% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ferx.R
\name{ferx_write_data}
\alias{ferx_write_data}
\title{Write a data frame as CSV for ferx input}
\usage{
ferx_write_data(data, path)
}
\arguments{
\item{data}{Data frame to write.}

\item{path}{Output file path.}
}
\description{
Write a data frame as CSV for ferx input
}
\keyword{internal}
