% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removal_metrics.R
\name{load_removal_results}
\alias{load_removal_results}
\title{Load results from a previous `remove_erroneous_obs()` run}
\usage{
load_removal_results(run_id)
}
\arguments{
\item{run_id}{Character. The `run_id` used in [remove_erroneous_obs()].}
}
\value{
The results list as saved by [remove_erroneous_obs()].
}
\description{
Load results from a previous `remove_erroneous_obs()` run
}
