% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nm.R
\name{nm_write_model}
\alias{nm_write_model}
\title{Write a NONMEM model object to a `.mod` file}
\usage{
nm_write_model(model = NULL, modelfile = NULL, overwrite = FALSE)
}
\arguments{
\item{model}{A NONMEM model object created by [nm_read_model()].}

\item{modelfile}{Output file path.}

\item{overwrite}{Overwrite an existing file?  Default `FALSE`.}
}
\description{
Write a NONMEM model object to a `.mod` file
}
\keyword{internal}
