% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_cleaning.R
\name{plot.irxclean_quality}
\alias{plot.irxclean_quality}
\title{Plot method for irxclean_quality objects}
\usage{
\method{plot}{irxclean_quality}(x, ...)
}
\arguments{
\item{x}{An `irxclean_quality` object from [assess_data_quality()].}

\item{...}{Ignored.}
}
\value{
A `ggplot` object (or a list of `ggplot` objects).
}
\description{
Generates a panel of up to three diagnostic plots:
covariate distributions with outlier bounds, flagged concentration profiles,
and missing-data rates.
}
