% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.R
\name{plot.irxclean_stability}
\alias{plot.irxclean_stability}
\title{Plot method for irxclean_stability objects}
\usage{
\method{plot}{irxclean_stability}(x, ...)
}
\arguments{
\item{x}{An `irxclean_stability` object from [check_model_stability()].}

\item{...}{Ignored.}
}
\value{
A list of \code{ggplot} objects: \code{$parameter_drift} and
  \code{$param_rse}.
}
\description{
Produces two panels:
\enumerate{
  \item Parameter trajectories (\% change from baseline) coloured by
    stability status (stable = teal, drifted = grey, unstable = red).
  \item RSE bar chart per parameter, sorted by RSE and coloured by status.
}
}
