% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_plots.R
\name{print.irxclean_demographics}
\alias{print.irxclean_demographics}
\title{Print method for irxclean_demographics objects}
\usage{
\method{print}{irxclean_demographics}(x, ...)
}
\arguments{
\item{x}{An `irxclean_demographics` object.}

\item{...}{Ignored.}
}
\description{
Print method for irxclean_demographics objects
}
