% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusion_criteria.R
\name{print.irxclean_exclusions}
\alias{print.irxclean_exclusions}
\title{Print method for irxclean_exclusions objects}
\usage{
\method{print}{irxclean_exclusions}(x, ...)
}
\arguments{
\item{x}{An \code{irxclean_exclusions} object.}

\item{...}{Ignored.}
}
\description{
Print method for irxclean_exclusions objects
}
