% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_cleaning.R
\name{print.irxclean_quality}
\alias{print.irxclean_quality}
\title{Print method for irxclean_quality objects}
\usage{
\method{print}{irxclean_quality}(x, ...)
}
\arguments{
\item{x}{An `irxclean_quality` object.}

\item{...}{Ignored.}
}
\description{
Print method for irxclean_quality objects
}
