% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.R
\name{print.irxclean_stability}
\alias{print.irxclean_stability}
\title{Print method for irxclean_stability objects}
\usage{
\method{print}{irxclean_stability}(x, ...)
}
\arguments{
\item{x}{An `irxclean_stability` object.}

\item{...}{Ignored.}
}
\description{
Print method for irxclean_stability objects
}
