% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.R
\name{stability_flag_table}
\alias{stability_flag_table}
\title{Summarise unstable parameters in a tidy table}
\usage{
stability_flag_table(stability)
}
\arguments{
\item{stability}{An `irxclean_stability` object.}
}
\value{
A data frame of unstable parameters with stability metrics.
}
\description{
Returns the rows of the parameter summary for parameters classified as
\code{"unstable"} (i.e. those that did not reach a stable estimate).
}
